# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Corbin's Project and Task Management API",
    "version": "1.0.0",
    "description": "A REST API built with Node.js, Express, MongoDB, Mongoose and Google OAuth. The API provides CRUD operations for projects and tasks."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://YOUR-RENDER-APP.onrender.com",
      "description": "Production Render server"
    }
  ],
...
  "paths": {
...
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Google OAuth callback",
        "description": "Handles Google's OAuth callback after successful authentication.",
        "responses": {
          "302": {
            "description": "Redirects to Swagger documentation after successful login."
          },
          "401": {
            "description": "Authentication failed."
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current authenticated user",
        "responses": {
          "200": {
            "description": "Current user returned successfully."
          },
          "401": {
            "description": "User is not authenticated."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
...
  }
}